+++ VTIL-Architecture/symex/variable.cpp
@@ -188,7 +188,7 @@ namespace vtil::symbolic
 		{
 			// Iterate each operand:
 			//
-			for ( int i = 0; i < it->base->operand_count(); i++ )
+			for ( size_t i = 0; i < it->base->operand_count(); i++ )
 			{
 				// Skip if not register.
 				//
