+++ VTIL-Compiler/optimizer/istack_ref_substitution_pass.cpp
@@ -54,7 +54,7 @@ namespace vtil::optimizer
 
 				// If successful, replace the operands.
 				//
-				if ( auto stack_offset = delta.get<int64_t>() )
+				if ( auto stack_offset = delta.get<intptr_t>() )
 				{
 					( +it )->operands[ it->base->memory_operand_index ] = { REG_SP };
 					( +it )->operands[ it->base->memory_operand_index + 1 ].imm().ival += *stack_offset;
